refactor twilio backend token generation helpers into functions/src/utils/twilio.js
