+++ functions/video.js
@@ -1,27 +1,8 @@
 const functions = require("firebase-functions");
 
-const twilio = require("twilio");
-const AccessToken = twilio.jwt.AccessToken;
-const { VideoGrant } = AccessToken;
+const { twilioVideoToken } = require("./src/utils/twilio");
 
 const PROJECT_ID = functions.config().project.id;
-const TWILIO_CONFIG = functions.config().twilio;
-
-const generateTwilioToken = () => {
-  return new AccessToken(
-    TWILIO_CONFIG.account_sid,
-    TWILIO_CONFIG.api_key,
-    TWILIO_CONFIG.api_secret
-  );
-};
-
-const twilioVideoToken = (identity, room) => {
-  const videoGrant = new VideoGrant({ room });
-  const token = generateTwilioToken();
-  token.addGrant(videoGrant);
-  token.identity = identity;
-  return token;
-};
 
 exports.getTwilioToken = functions.https.onCall((data, context) => {
   if (!context.auth || !context.auth.token) {
